Fix roon API test app for latest browser WebSocket and Roon API changes
* Add polyfills for the native WebSocket browser implementation to be compatible with nodejs ws methods that the Roon API library uses
* Add periodic reconnect (every 8 seconds) to get around disconnect every 10 seconds due to inability of browser WebSocket to send a Ping often enough. Unfortunately there is not a solution to this without changing the Roon API behavior.

Works fine with these changes and is actually a great app for Roon; however the view flashes every 8 seconds.

See issue https://github.com/RoonLabs/node-roon-api/issues/34 for core issue with roon js API using non-native WebSocket methods. Though, it is billed as a "node roon API" so not sure if that's in scope.